+++ src/api/index.ts
@@ -1,6 +1,5 @@
 export { hashZomeCall } from "@holochain/serialization";
 export * from "./admin/index.js";
-export * from "./app-agent/index.js";
 export * from "./app/index.js";
 export { IsoWebSocket, WsClient } from "./client.js";
 export {